+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Run the system test on Travis (#113)
 - Add security headers to the default response (#110)
 - Whitelist BadRequestHttpException so that messages are not sanitized (#108)
+- REST API endpoint for adding a subscriber to a list
 
 ### Changed